feat(canvas): add drawing functionality to Rainbow MVC app
Extend the Rainbow MVC application by adding a canvas widget that allows users to draw with selected colors. The drawing color is updated when a color button is clicked. The canvas supports smooth line drawing with adjustable color and width. The window size has been increased to accommodate the new canvas widget.